+++ config-sample.yaml
@@ -5,13 +5,4 @@ server: 127.0.0.1
 port: 8050
 support_email: help@iri.columbia.edu
 
-data_dir: data/
-prefix: /python_maproom
-tile_pfx: /python_maproom/tile
-map_title: Monthly Climatology Maproom
-
-vars:
-  Rainfall: rfe
-  Minimum Temperature: tmin
-  Maximum Temperature: tmax
-  Mean Temperature: tmean
+prefix: /my_first_maproom
